resolve ocean-area circular dependency
