+++ packages/ocean-areas/src/data/geometries.ts
@@ -1,5 +1,5 @@
 import { FeatureCollection, Geometry } from 'geojson'
-import { OceanAreaProperties } from '..'
+import { OceanAreaProperties } from '../ocean-areas'
 
 const oceanAreas: FeatureCollection<Geometry, OceanAreaProperties> = {
   type: 'FeatureCollection',
